Graph.java and VariableDependency.java were moved to graphData package.
